collection, access_indicators, resource_url.
